Added in gatekeeper middleware isAdminOrSelf to users and pastOrders, deleted password from user put route return data, created utils file for gatekeeper middleware
